Run code several times and log how many times it fails

+++ fluid/compile.sh
@@ -2,7 +2,9 @@
 
 CONFIG="${1:-release}"
 
-rm -r build/bin/{Fluid,Tests}
+if [ -d "build/bin/" ]; then
+    rm -r build/bin/{Fluid,Tests}
+fi
 
 if [ "premake5.lua" -nt "generated/Makefile" ]; then
     echo "Newer premake, deleting generated"
@@ -19,8 +21,18 @@ pushd "generated"
     SUCCESS=$?
 popd
 
+FAILED=0
+
 if [ $SUCCESS -eq 0 ]; then
-    ./build/bin/Fluid/$CONFIG/Fluid
+    for i in {1..1}
+    do
+        echo "Run: $i"
+        timeout 2s ./build/bin/Fluid/$CONFIG/Fluid
+        if [ $? -ne 124 ]; then
+            ((FAILED++))
+        fi
+        echo "Failed runs: $FAILED / $i"
+    done
 else
     echo "Build failed"
 fi
